refactor(front): rename JobStatus{Label→Badge}
