+++ frontend/src/components/job/JobStatusBadge.vue
